+++ rust_dev_preview/dynamodb/src/scenario/add.rs
@@ -1,0 +1,52 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+use aws_sdk_dynamodb::model::AttributeValue;
+use aws_sdk_dynamodb::{Client, Error};
+
+pub struct Item {
+    pub p_type: String,
+    pub age: String,
+    pub username: String,
+    pub first: String,
+    pub last: String,
+}
+
+// Add an item to a table.
+// snippet-start:[dynamodb.rust.add-item]
+pub async fn add_item(client: &Client, item: Item, table: &String) -> Result<(), Error> {
+    let user_av = AttributeValue::S(item.username);
+    let type_av = AttributeValue::S(item.p_type);
+    let age_av = AttributeValue::S(item.age);
+    let first_av = AttributeValue::S(item.first);
+    let last_av = AttributeValue::S(item.last);
+
+    let request = client
+        .put_item()
+        .table_name(table)
+        .item("username", user_av)
+        .item("account_type", type_av)
+        .item("age", age_av)
+        .item("first_name", first_av)
+        .item("last_name", last_av);
+
+    println!("Executing request [{request:?}] to add item...");
+
+    let resp = request.send().await?;
+
+    let attributes = resp.attributes().unwrap();
+
+    println!(
+        "Added user {:?}, {:?} {:?}, age {:?} as {:?} user",
+        attributes.get("username"),
+        attributes.get("first_name"),
+        attributes.get("last_name"),
+        attributes.get("age"),
+        attributes.get("p_type")
+    );
+
+    Ok(())
+}
+// snippet-end:[dynamodb.rust.add-item]
