+++ rust_dev_preview/dynamodb/src/scenario/mod.rs
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+pub mod add;
 pub mod create;
 pub mod delete;
 pub mod list;
